Add functionality  for using internal pullups with i2c into the Wire lib.
Add the variable  usePullups to have a simple Bool for switching the usage of the pullups.
Default that to True for legacy compatibility reasons. 

Add Methods for disabling, enabling and getting the current status of the pullups.
Note: Since those methods only affect the variable, they can only affect the usage before Wire.begin() was called.